have the child processes clean up their own state if possible

+++ lib/trident/worker.rb
@@ -13,7 +13,7 @@ module Trident
     # Remove the pidfile associated with this
     # worker
     def destroy
-      FileUtils.rm path
+      FileUtils.rm path if File.exists?(path)
     end
 
     # We determine the time that this worker was
